flex/requirements.txt
Flask==0.12.1
gunicorn==19.7.1
pylibmc==1.5.2
google-cloud-datastore==0.22.1